update compression script

+++ arithmetization/src/test/resources/compress.sh
@@ -18,28 +18,28 @@ if [ ! -d "$TARGET_DIR" ]; then
   exit 1
 fi
 
-# Function to compress and rename a single file
-compress_and_rename() {
-  file="$1"
-  TARGET_DIR="$2"
-  original_filename=$(basename "$file")
-  base_name=$(echo "$original_filename" | grep -o '^[0-9]\+-[0-9]\+')
-  temp_filename="${base_name}.tmp.gz"
-  new_filename="${base_name}.conflated.v0.2.0.lt.gz"
-  echo "Compressing $file to $TARGET_DIR/$temp_filename..."
-  pigz -c "$file" > "$TARGET_DIR/$temp_filename"
-  if [ $? -eq 0 ]; then
-    mv "$TARGET_DIR/$temp_filename" "$TARGET_DIR/$new_filename"
-    echo "Renamed $TARGET_DIR/$temp_filename to $TARGET_DIR/$new_filename"
-  else
-    echo "Compression failed for $file"
-    rm -f "$TARGET_DIR/$temp_filename"
+# Get the total number of files to compress
+total_files=$(find "$TARGET_DIR" -type f | wc -l)
+index=1
+
+# Compress and rename each file in the target directory
+for file in "$TARGET_DIR"/*; do
+  if [ -f "$file" ]; then
+    original_filename=$(basename "$file")
+    base_name=$(echo "$original_filename" | grep -o '^[0-9]\+-[0-9]\+')
+    temp_filename="${base_name}.tmp.gz"
+    new_filename="${base_name}.conflated.v0.2.0.lt.gz"
+    echo "Compressing $file to $TARGET_DIR/$temp_filename... ($index/$total_files)"
+    pigz -1 -p "$NUM_CPUS" -c "$file" > "$TARGET_DIR/$temp_filename"
+    if [ $? -eq 0 ]; then
+      mv "$TARGET_DIR/$temp_filename" "$TARGET_DIR/$new_filename"
+      echo "Renamed $TARGET_DIR/$temp_filename to $TARGET_DIR/$new_filename"
+    else
+      echo "Compression failed for $file"
+      rm -f "$TARGET_DIR/$temp_filename"
+    fi
+    index=$((index + 1))
   fi
-}
-
-export -f compress_and_rename
-
-# Find all files and compress them in parallel
-find "$TARGET_DIR" -type f | xargs -I {} -P "$NUM_CPUS" bash -c 'compress_and_rename "{}" "$0"' "$TARGET_DIR"
+done
 
 echo "Compression and renaming completed."
